refactor(mvc): move map types into instruction namespace
State.Map and its inner types now live on the map instruction itself,
matching how ref, get, set, and def declare theirs: map.Keyed replaces
State.Map, map.Create replaces State.Map.Factory, and map.Key moves
alongside them.
