+++ skills/SKILL.md
@@ -133,9 +133,9 @@ class UserProfile extends State {
 | `map(input => value)` | `add(input?)` spawns; a string input is also the key (occupied throws), otherwise keyed by `String(value)`. `set(key)` alone respawns. |
 | `map(StateClass)`   | `add(input?)` instantiates the class; object input assigns fields, string input keys the entry and becomes a `Component`'s `key`. |
 
-`map()` is a field instruction: it resolves into a `State.Map<K, V>` (extending native `Map`) when the hosting state activates, and is not usable standalone. The map has reactive `get(key)`, `has(key)`, `size`, iteration, `set`, `delete`, and `clear`. Calling `get()` with no key returns a shallow `ReadonlyMap` snapshot. `keys(fn)` / `values(fn)` / `entries(fn)` return reusable iterables of transformed results (`throw false` skips an entry), tracking like their plain forms.
+`map()` is a field instruction: it resolves into a `map.Keyed<K, V>` (extending native `Map`) when the hosting state activates, and is not usable standalone. The map has reactive `get(key)`, `has(key)`, `size`, iteration, `set`, `delete`, and `clear`. Calling `get()` with no key returns a shallow `ReadonlyMap` snapshot. `keys(fn)` / `values(fn)` / `entries(fn)` return reusable iterables of transformed results (`throw false` skips an entry), tracking like their plain forms.
 
-With a factory, the map is string-keyed (`State.Map.Factory<V>`) and owns what `add` spawns: spawned `State` values are destroyed when deleted, cleared, or replaced. Activated values supplied directly via `set(key, value)` are guests and never destroyed.
+With a factory, the map is string-keyed (`map.Create<V>`) and owns what `add` spawns: spawned `State` values are destroyed when deleted, cleared, or replaced. Activated values supplied directly via `set(key, value)` are guests and never destroyed.
 
 Every map is adopted by its hosting state when the instruction resolves at activation; the field is read-only. Fresh (never-activated) members - spawned, stored, or present at adoption - are parented to the owner, activate inside its context, and are destroyed with it; already-activated values keep guest status. A `State` value that dies evicts itself from every key it occupies.
